Changed price type

+++ src/business/excursion/SitesActiv.java
@@ -1,11 +1,13 @@
 package business.excursion;
 
+import java.math.BigDecimal;
+
 public class SitesActiv extends SiteTouristique {
 	
 	private float duration;
 	
 	@Override
-	public double getPrix() {
+	public BigDecimal getPrix() {
 		return super.getPrix();
 	}
 
@@ -17,7 +19,7 @@ public class SitesActiv extends SiteTouristique {
 		return duration;
 	}
 
-	public SitesActiv(String nom, String description, Adresse adresse, double prix, int id, float duration) {
+	public SitesActiv(String nom, String description, Adresse adresse, BigDecimal prix, int id, float duration) {
 		super(nom, description, adresse, prix, id);
 		this.duration = duration;
 	}
